Fixing redux store and apollo client circular dependencies
