+++ src/components/Navbar/InsightsDropdown.js
@@ -6,7 +6,7 @@ import cx from 'classnames'
 import Icon from '@santiment-network/ui/Icon'
 import Button from '@santiment-network/ui/Button'
 import { InputWithIcon as Input } from '@santiment-network/ui/Input'
-import { store } from '../../index'
+import { store } from '../../redux'
 import { showNotification } from '../../actions/rootActions'
 import GA from './../../utils/tracking'
 import { checkIsLoggedIn } from '../../pages/UserSelectors'
