+++ src/components/SubscriptionForm/SubscriptionForm.js
@@ -8,7 +8,7 @@ import Button from '@santiment-network/ui/Button'
 import Input from '@santiment-network/ui/Input'
 import { Checkbox } from '@santiment-network/ui/Checkboxes'
 import { EMAIL_LOGIN_MUTATION } from './loginGQL'
-import { store } from '../../index'
+import { store } from '../../redux'
 import { showNotification } from '../../actions/rootActions'
 import GA from './../../utils/tracking'
 import styles from './SubscriptionForm.module.scss'
