+++ src/components/TimelineEventComments/TimelineEventComments.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Comments from '../Insight/comments/Comments'
-import { client } from '../../index'
+import { client } from '../../apollo'
 import {
   COMMENTS_TIMELINE_EVENTS_QUERY,
   CREATE_TIMELINE_EVENT_COMMENT_MUTATION
