+++ src/components/Trends/TrendsTable/TrendsTable.js
@@ -14,7 +14,7 @@ import Label from '@santiment-network/ui/Label'
 import Icon from '@santiment-network/ui/Icon'
 import Tooltip from '@santiment-network/ui/Tooltip'
 import Button from '@santiment-network/ui/Button'
-import { store } from '../../../index'
+import { store } from '../../../redux'
 import ValueChange from '../../../components/ValueChange/ValueChange'
 import WordCloud from '../../../components/WordCloud/WordCloud'
 import InsightCardSmall from '../../../components/Insight/InsightCardSmall'
