+++ src/ducks/Studio/Chart/PaywallInfo.js
@@ -3,7 +3,7 @@ import gql from 'graphql-tag'
 import Button from '@santiment-network/ui/Button'
 import Icon from '@santiment-network/ui/Icon'
 import Tooltip from '@santiment-network/ui/Tooltip'
-import { client } from '../../../index'
+import { client } from '../../../apollo'
 import { getDateFormats } from '../../../utils/dates'
 import UpgradeBtn from '../../../components/UpgradeBtn/UpgradeBtn'
 import { useUserSubscriptionStatus } from '../../../stores/user/subscriptions'
