+++ src/ducks/Studio/Chart/index.js
@@ -13,6 +13,7 @@ import ChartFullscreenBtn from './Fullscreen'
 import Insights from './Insights'
 import Compare from '../Compare'
 import { extractMirrorMetricsDomainGroups } from '../utils'
+import { useMetricColor } from '../Widget/ChartWidgetColorProvider'
 import { useAllTimeData } from '../timeseries/hooks'
 import Chart from '../../Chart'
 import Signals from '../../Chart/Signals'
@@ -20,7 +21,6 @@ import { useMetricCategories } from '../../Chart/Synchronizer'
 import { useDomainGroups, useAxesMetricsKey } from '../../Chart/hooks'
 import { useHighlightMetricColor } from '../../Chart/colors'
 import { useUser } from '../../../stores/user'
-import { useMetricColor } from '../Widget/ChartWidgetColorProvider'
 import styles from './index.module.scss'
 
 const Canvas = ({
