+++ src/ducks/Studio/timeseries/hooks.js
@@ -1,10 +1,10 @@
 import { useState, useEffect } from 'react'
-import { client } from '../../../index'
 import { getQuery, getPreTransform } from './fetcher'
 import { normalizeDatetimes, mergeTimeseries } from './utils'
 import { substituteErrorMsg } from './errors'
-import { getIntervalByTimeRange } from '../../../utils/dates'
 import { getAvailableInterval } from '../../dataHub/metrics/intervals'
+import { getIntervalByTimeRange } from '../../../utils/dates'
+import { client } from '../../../apollo'
 
 // NOTE: Polyfill for a PingdomBot 0.8.5 browser (/sentry/sanbase-frontend/issues/29459/) [@vanguard | Feb 6, 2020]
 window.AbortController =
