+++ src/pages/Account/SettingsNotifications.js
@@ -8,7 +8,7 @@ import Label from '@santiment-network/ui/Label'
 import Selector from '@santiment-network/ui/Selector/Selector'
 import Settings from './Settings'
 import * as actions from '../../actions/types'
-import { store } from '../../index'
+import { store } from '../../redux'
 import { showNotification } from '../../actions/rootActions'
 import SettingsTelegramNotifications from './SettingsTelegramNotifications'
 import SettingsEmailNotifications from './SettingsEmailNotifications'
