+++ src/stores/user/subscriptions.js
@@ -2,7 +2,7 @@ import { useMemo } from 'react'
 import gql from 'graphql-tag'
 import { useQuery } from '@apollo/react-hooks'
 import { buildRefetcher, update } from './utils'
-import { client } from '../../index'
+import { client } from '../../apollo'
 import {
   PLANS,
   getSanbaseSubscription,
